+++ flutter/packages/growerp_models/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.1
+* reset password fixed
+* product conversion changes
+
 ## 1.2.0
 * Now centrally all models for flutter and terminal.